+++ app/controllers/member/rates_controller.rb
@@ -5,16 +5,11 @@ class Member::RatesController < Member::BaseController
     id = params[:id]
     rateable = type.constantize.find(id)
     
-    if(!rateable.ratings.include?(Rating.find_by_user_id(current_user.id)))
-      rating = rateable.rate params[:rate]
-      rating.user_id = current_user.id
-      if rating.save
-        flash[:ok] = I18n.t("tog_core.site.rating.added")
-      else
-        flash[:error] = I18n.t("tog_core.site.rating.error")
-      end
-    else
+    if rateable.rated_by?(current_user) == true
       flash[:error] = I18n.t("tog_core.site.rating.already_rated")
+    else
+      rateable.rate_it params[:rate], current_user
+      flash[:ok] = I18n.t("tog_core.site.rating.added")
     end
     respond_to do |format|
       format.html { redirect_to request.referer }