+++ db/migrate/007_remove_user_from_ratings.rb
@@ -1,0 +1,11 @@
+class RemoveUserFromRatings < ActiveRecord::Migration
+
+  def self.up
+    remove_column :ratings, :user_id
+  end
+
+  def self.down
+    add_column :ratings, :user_id, :integer
+  end
+  
+end