+++ init.rb
@@ -1,7 +1,6 @@
 require_plugin 'acts_as_scribe'
 require_plugin 'acts_as_taggable_on_steroids'
 require_plugin 'acts_as_abusable'
-require_plugin 'acts_as_rateable'
 require_plugin 'viking'
 
 # require the will_paginate as a gem. This could be ´config.gem´ as well.